Gate APIs behind a build-time flag
